Add table-based formatter
E.g.,

    $ syn < file.txt | head -5
    11 8 Noun families
    61 4 Noun ones
    112 6 Noun writer
    126 9 Noun beginning
    148 5 Noun novel

The format is "<location> <length> <tag> <token>".

I've renamed the "term" formatter to "ansi" in the process, which is
more true to what it is.

I considered making "table" the default (or checking
isatty(fileno(stdout)) to decide whether to set -fansi or -ftable, but
it seemed a bit "magical" (and made the results of things like piping to
less less-than-obvious), and the best demo is the demo that works out of
the box: and it's easy for tools to explicitly set the formatter to what
they need it to be.

I also cleaned up JSON formatter by normalizing/passing tokens to the
processor directly.

Signed-off-by: Stephen Celis <stephen@stephencelis.com>